docs: consolidate roadmap, changelog, and documentation standards
Roadmap:

- Merged ROADMAP.md + ROADMAP-updates.md + PHASE_10_0_IMPLEMENTATION.md into single ROADMAP.md

- Phase 10.0 marked complete (tokens, components, BookCard v4, ThreeDotMenu, warm gradients)

- Phase 10.0C full conversion plan added (8 sessions, 62 files, 10 work groups)

- Critical fixes cross-referenced to 10.0C groups

- Rating float bug documented in Technical Debt as gotcha

- Version bumped to v0.34.0

Changelog:

- Merged CHANGELOG-entry.md into single CHANGELOG.md

- Full v0.31.0 entry restored (nav redesign C1-C3, was skeleton in other versions)

- v0.34.0 entry added (warm gradients, BookCard v4, ThreeDotMenu extraction, rating bug fix)

- Complete history v0.20.0–v0.34.0, no detail loss

Documentation standards:

- Added Golden Rule #5 (single-file documentation) to cursorrules

- Updated .cursorrules with current file structure, design tokens, protected systems, naming conventions

- Killed:  ROADMAP patch files, CHANGELOG fragments

+++ backend/routers/titles.py
@@ -66,7 +66,7 @@ class TitleSummary(BaseModel):
     series_number: Optional[str] = None
     category: Optional[str] = None
     status: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     word_count: Optional[int] = None
     cover_gradient: Optional[str] = None
     cover_bg_color: Optional[str] = None
@@ -92,7 +92,7 @@ class TitleDetail(BaseModel):
     series_number: Optional[str] = None
     category: Optional[str] = None
     status: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
     date_started: Optional[str] = None
     date_finished: Optional[str] = None
     publication_year: Optional[int] = None
@@ -167,7 +167,7 @@ class SeriesBookItem(BaseModel):
     title: str
     series_number: Optional[str] = None
     status: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
 
 
 class SeriesDetail(BaseModel):
@@ -225,7 +225,7 @@ class BookStatusUpdate(BaseModel):
 
 class BookRatingUpdate(BaseModel):
     """Request body for updating a title's rating."""
-    rating: Optional[int] = None  # None to clear rating, 1-5 to set
+    rating: Optional[float] = None  # None to clear rating, 1-5 to set
 
 
 class BookDatesUpdate(BaseModel):
